19w36a

+++ minecraft/src/net/minecraft/world/entity/monster/MagmaCube.java
@@ -114,8 +114,8 @@ public class MagmaCube extends Slime {
 	}
 
 	@Override
-	protected int getAttackDamage() {
-		return super.getAttackDamage() + 2;
+	protected float getAttackDamage() {
+		return super.getAttackDamage() + 2.0F;
 	}
 
 	@Override
